activity loading for deck layout the responsibility of the deck
