+++ assets/src/apps/delivery/store/features/activities/slice.ts
@@ -1,5 +1,4 @@
 import {
-  createAsyncThunk,
   createEntityAdapter,
   createSelector,
   createSlice,
@@ -8,11 +7,8 @@ import {
   PayloadAction,
   Slice,
 } from '@reduxjs/toolkit';
-import { getActivityForDelivery, getBulkActivitiesForAuthoring } from 'data/persistence/activity';
-import { getBulkAttemptState } from 'data/persistence/state/intrinsic';
-import { ResourceId } from 'data/types';
 import { RootState } from '../../rootReducer';
-import { loadPageState, selectSectionSlug, selectSequence } from '../page/slice';
+import { loadPageState } from '../page/slice';
 
 interface IActivity {
   // TODO
@@ -71,66 +67,6 @@ export const {
   deleteActivities,
   setCurrentActivityId,
 } = slice.actions;
-
-export const fetchActivity = createAsyncThunk(
-  `${ActivitiesSlice}/fetchActivity`,
-  async (activityId: ResourceId, thunkApi) => {
-    const sectionSlug = selectSectionSlug(thunkApi.getState() as RootState);
-    const activity = await getActivityForDelivery(sectionSlug, activityId);
-    // TODO: need a sequence ID and/or some other ID than db id to use here
-    thunkApi.dispatch(upsertActivity({ activity: { ...activity, id: activityId.toString() } }));
-  },
-);
-
-export const loadActivities = createAsyncThunk(
-  `${ActivitiesSlice}/loadActivities`,
-  async (activityIds: ResourceId[], thunkApi) => {
-    const sectionSlug = selectSectionSlug(thunkApi.getState() as RootState);
-    const results = await getBulkActivitiesForAuthoring(sectionSlug, activityIds);
-    const sequence = selectSequence(thunkApi.getState() as RootState);
-    const activities = results.map((result) => {
-      const sequenceEntry = sequence.find((entry: any) => entry.activity_id === result.id);
-      if (!sequenceEntry) {
-        console.warn(`Activity ${result.id} not found in the page model!`);
-        return;
-      }
-      const activity = {
-        id: sequenceEntry.custom.sequenceId,
-        resourceId: sequenceEntry.activity_id,
-        content: result.content,
-      };
-      return activity;
-    });
-    // TODO: need a sequence ID and/or some other ID than db id to use here
-    thunkApi.dispatch(setActivities({ activities }));
-  },
-);
-
-export const loadActivityState = createAsyncThunk(
-  `${ActivitiesSlice}/loadActivityState`,
-  async (attemptGuids: string[], thunkApi) => {
-    const sectionSlug = selectSectionSlug(thunkApi.getState() as RootState);
-    const results = await getBulkAttemptState(sectionSlug, attemptGuids);
-
-    // TODO: map back to activities in model and update everything
-    const sequence = selectSequence(thunkApi.getState() as RootState);
-    const activities = results.map((result) => {
-      const sequenceEntry = sequence.find((entry: any) => entry.activity_id === result.activityId);
-      if (!sequenceEntry) {
-        console.warn(`Activity ${result.activityId} not found in the page model!`);
-        return;
-      }
-      const activity = {
-        id: sequenceEntry.custom.sequenceId,
-        resourceId: sequenceEntry.activity_id,
-        content: result.model,
-      };
-      return activity;
-    });
-
-    thunkApi.dispatch(setActivities({ activities }));
-  },
-);
 
 // SELECTORS
 export const selectState = (state: RootState): ActivitiesState =>
